pyproject.toml
[project]
name = "pizza-voice-agent"
version = "0.1.0"
description = "LangGraph-driven voice agent that places pizza orders via outbound call."
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "websockets>=12.0",
  "twilio>=9.0",
  "deepgram-sdk>=3.4",
  "cartesia>=1.0",
  "silero-vad>=5.0",
  "torch>=2.2",
  "langgraph>=0.2",
  "langchain-core>=0.3",
  "langchain-openai>=0.2",
  "langchain-groq>=0.2",
  "langfuse>=2.36",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "scipy>=1.12",
  "numpy>=1.26",
  "pyngrok>=7.0",
  "httpx>=0.27",
  "audioop-lts>=0.2.2",
  "pipecat-ai[cartesia,deepgram]==0.0.108",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-mock>=3.12",
  "ruff>=0.4",
  "pyright>=1.1.350",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "ANN", "D"]
ignore = ["D203", "D213", "ANN101", "ANN102"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ANN"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
include = ["agent", "tests"]
strict = ["agent"]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
